+++ src/main/java/bestcommerce/brand/manager/controller/BrandController.java
@@ -5,8 +5,8 @@ import bestcommerce.brand.manager.entity.Manager;
 import bestcommerce.brand.manager.service.BrandService;
 import bestcommerce.brand.manager.service.ManagerService;
 import bestcommerce.brand.manager.service.RoleService;
-import bestcommerce.brand.util.EntityConverter;
-import bestcommerce.brand.util.ManagerRole;
+import bestcommerce.brand.util.converter.EntityConverter;
+import bestcommerce.brand.manager.ManagerRole;
 import bestcommerce.brand.util.ResponseDto;
 import bestcommerce.brand.util.ResponseStatus;
 import bestcommerce.brand.util.exception.DuplicateBrandManagerException;
