+++ src/main/java/bestcommerce/brand/product/controller/ProductController.java
@@ -12,8 +12,8 @@ import bestcommerce.brand.product.service.ProductService;
 import bestcommerce.brand.size.dto.SizeDto;
 import bestcommerce.brand.size.service.QuantityService;
 import bestcommerce.brand.size.service.SizeService;
-import bestcommerce.brand.util.DtoConverter;
-import bestcommerce.brand.util.EntityConverter;
+import bestcommerce.brand.util.converter.DtoConverter;
+import bestcommerce.brand.util.converter.EntityConverter;
 import bestcommerce.brand.util.ResponseDto;
 import bestcommerce.brand.util.ResponseStatus;
 import lombok.RequiredArgsConstructor;
